@@ -8,6 +8,8 @@ import snap.domains.photographer.entity.PhotographerSchedule;
 import snap.domains.photographer.repository.PhotographerScheduleRepository;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -17,14 +19,27 @@ import java.util.stream.Collectors;
 public class PhotographerScheduleDomainService {
     private final PhotographerScheduleRepository scheduleRepository;
 
-    public void updateSchedule(Photographer photographer, List<LocalDateTime> unableDates){
-        unableDates.stream()
-                .map(unableDate -> scheduleRepository.save(
-                        PhotographerSchedule.builder()
-                                .photographer(photographer)
-                                .unableDate(unableDate)
-                                .build()
-                ))
-                .collect(Collectors.toList());
+    public void updateSchedule(Photographer photographer, List<LocalDateTime> unableDates) {
+        List<LocalDateTime> existedUnableDates = photographer.getUnableSchedules()
+                .stream().map(PhotographerSchedule::getUnableDate).collect(Collectors.toList());
+
+        if (!(new HashSet<>(unableDates).containsAll(existedUnableDates) && new HashSet<>(existedUnableDates).containsAll(unableDates))) {
+            List<PhotographerSchedule> addList = new ArrayList<>();
+            for (LocalDateTime dateTime : unableDates) {
+                if (!existedUnableDates.contains(dateTime)) {
+                    addList.add(PhotographerSchedule.builder().photographer(photographer).unableDate(dateTime).build());
+                }
+            }
+
+            List<PhotographerSchedule> removeList = new ArrayList<>();
+            for (PhotographerSchedule dateTime : photographer.getUnableSchedules()) {
+                if (!unableDates.contains(dateTime.getUnableDate())) {
+                    removeList.add(dateTime);
+                }
+            }
+
+            scheduleRepository.saveAllAndFlush(addList);
+            scheduleRepository.deleteAll(removeList);
+        }
     }
 }
