@@ -5,9 +5,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import snap.domains.photographer.entity.Photographer;
 import snap.domains.photographer.entity.Special;
-import snap.domains.photographer.entity.SpecialKeyword;
 import snap.domains.photographer.repository.SpecialRepository;
+import snap.enums.SpecialKeyword;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -17,14 +18,25 @@ import java.util.stream.Collectors;
 public class SpecialDomainService {
     private final SpecialRepository specialRepository;
 
-    public void updateSpecial(Photographer photographer, List<SpecialKeyword> specialList) {
-        specialList.stream()
-                .map(keyword -> specialRepository.save(
-                        Special.builder()
-                                .photographer(photographer)
-                                .keyword(keyword)
-                                .build()
-                ))
-                .collect(Collectors.toList());
+    public void updateSpecial(Photographer photographer, List<SpecialKeyword> keywordList) {
+        List<SpecialKeyword> existedSpecialKeyword = photographer.getSpecialList().stream().map(Special::getKeyword).collect(Collectors.toList());
+
+
+        List<Special> addList = new ArrayList<>();
+        for (SpecialKeyword keyword : keywordList) {
+            if (!existedSpecialKeyword.contains(keyword)) {
+                addList.add(Special.builder().photographer(photographer).keyword(keyword).build());
+            }
+        }
+
+        List<Special> removeList = new ArrayList<>();
+        for (Special special : photographer.getSpecialList()) {
+            if (!keywordList.contains(special.getKeyword())) {
+                removeList.add(special);
+            }
+        }
+
+        specialRepository.saveAll(addList);
+        specialRepository.deleteAll(removeList);
     }
 }
