@@ -29,9 +29,11 @@ public class SecurityConfig {
                 .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
                 .and()
                 .authorizeRequests()
-                .antMatchers("/**").permitAll()
+                // .antMatchers("/**").permitAll()
                 .antMatchers("/members/test", "/members/signup", "/members/signin").permitAll()
                 .antMatchers("/photographers/**").permitAll()
+                .antMatchers("/plans/**").authenticated()
+                .antMatchers("/plans/deposit", "/plans/reserve").hasRole("PHOTOGRAPHER")
                 .antMatchers(HttpMethod.POST, "/members/**").authenticated()
                 .and()
                 .apply(new JwtSecurityConfig(jwtTokenUtil));
