@@ -1,16 +1,19 @@
 package snap.api.photographer.dto.request;
 
 import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import snap.domains.photographer.entity.PhotographerImage;
-import snap.domains.photographer.entity.SpecialKeyword;
+import snap.dto.request.ImageReq;
+import snap.dto.request.TagReq;
+import snap.enums.SpecialKeyword;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
 public class PhotographerCustomDto {
     private String nickname;
     private String profileImage;
@@ -21,17 +24,13 @@ public class PhotographerCustomDto {
 
     private List<Long> areaId;
 
-    private String instagram;
-    private String twitter;
-    private String kakaoChannel;
-    private String naverBlog;
-    private String homepage;
+    private SnsReq sns;
 
     private List<SpecialKeyword> specialList;
 
-    private List<String> tagList;
+    private TagReq tag;
 
     private List<LocalDateTime> unableDates;
 
-    private List<String> photographerImages;
+    private ImageReq image;
 }
