@@ -4,18 +4,21 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import snap.domains.photographer.entity.SpecialKeyword;
+import snap.enums.SpecialKeyword;
 import snap.domains.plan.entity.Plan;
 
 import java.time.LocalDateTime;
 
+/**
+ * 예약신청
+ */
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class PlanRequestDto {
     private Long photographerId;
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm")
+    // @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm")
     private LocalDateTime planDate;
-    private String category;
+    private SpecialKeyword category;
     private Long people;
     private String wishPlace;
     private String request;
@@ -23,7 +26,7 @@ public class PlanRequestDto {
     public Plan toEntity() {
         return Plan.builder()
                 .planDate(planDate)
-                .category(SpecialKeyword.valueOf(category))
+                .category(category)
                 .people(people)
                 .wishPlace(wishPlace)
                 .request(request)
