@@ -3,6 +3,7 @@ package snap.domains.photographer.entity;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import snap.dto.request.TagReq;
 
 import javax.persistence.*;
 
@@ -15,17 +16,27 @@ public class PhotographerTag {
     @Column(name = "photographer_tag_id")
     private Long photographerTagId;
 
-    @ManyToOne
+    @OneToOne
     @JoinColumn(name = "photographer_id")
     private Photographer photographer;
 
-    @ManyToOne
-    @JoinColumn(name = "tag_id")
-    private Tag tag;
+    @Column
+    private String tag1;
+
+    @Column
+    private String tag2;
+
+    @Column
+    private String tag3;
 
     @Builder
-    public PhotographerTag(Photographer photographer, Tag tag){
+    public PhotographerTag(Photographer photographer) {
         this.photographer = photographer;
-        this.tag = tag;
+    }
+
+    public void update(TagReq tag) {
+        this.tag1 = tag.getTag1();
+        this.tag2 = tag.getTag2();
+        this.tag3 = tag.getTag3();
     }
 }
