@@ -6,9 +6,9 @@ import org.springframework.transaction.annotation.Transactional;
 import snap.domains.photographer.entity.Photographer;
 import snap.domains.photographer.entity.PhotographerImage;
 import snap.domains.photographer.repository.PhotographerImageRepository;
+import snap.dto.request.ImageReq;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @Transactional
@@ -16,14 +16,13 @@ import java.util.stream.Collectors;
 public class PhotographerImageDomainService {
     private final PhotographerImageRepository imageRepository;
 
-    public void updatePhotographerImage(Photographer photographer, List<String> photographerImages){
-        photographerImages.stream()
-                .map(image -> imageRepository.save(
-                        PhotographerImage.builder()
-                                .photographer(photographer)
-                                .image(image)
-                                .build()
-                ))
-                .collect(Collectors.toList());
+    public void createPhotographer(Photographer photographer) {
+        imageRepository.save(
+                PhotographerImage.builder().photographer(photographer).build()
+        );
+    }
+
+    public void updatePhotographerImage(Photographer photographer, ImageReq image) {
+        photographer.getImages().update(image);
     }
 }
