@@ -1,6 +1,7 @@
 package snap.domains.plan.entity;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -8,7 +9,8 @@ import org.hibernate.annotations.GenericGenerator;
 import org.hibernate.annotations.Type;
 import snap.domains.member.entity.Member;
 import snap.domains.photographer.entity.Photographer;
-import snap.domains.photographer.entity.SpecialKeyword;
+import snap.enums.SpecialKeyword;
+import snap.enums.Status;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
@@ -16,7 +18,7 @@ import java.util.UUID;
 
 @Entity
 @Getter
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Plan {
     @Id
     @GeneratedValue(generator = "uuid2")
@@ -68,10 +70,14 @@ public class Plan {
     @Column
     private String request;
 
+    @Column
+    private String message;
+
     @Builder
-    public Plan(Member customer, Photographer photographer, Status status, LocalDateTime planDate,
+    public Plan(UUID planId, Member customer, Photographer photographer, Status status, LocalDateTime planDate,
                 SpecialKeyword category, Long people, Long price, String wishPlace, String placeName,
-                String placeAddress, Long x, Long y, String request) {
+                String placeAddress, Long x, Long y, String request, String message) {
+        this.planId = planId;
         this.customer = customer;
         this.photographer = photographer;
         this.status = status;
@@ -85,5 +91,20 @@ public class Plan {
         this.x = x;
         this.y = y;
         this.request = request;
+        this.message = message;
     }
+
+    public void detailPlan(UUID planId, Status status, Long price, String placeName, String placeAddress, String message) {
+        this.planId = planId;
+        this.status = status;
+        this.price = price;
+        this.placeName = placeName;
+        this.placeAddress = placeAddress;
+        this.message = message;
+    }
+
+    public void setStatus(Status status) {
+        this.status = status;
+    }
+
 }
